Dtos and Mappers for Products were added
If users send a GET request to the server to receive product(s) the server returns data in ProductReadingDto form.
If they send a POST request or a PUT request they have to use ProductCreationDto form.

There is some Mapper implementations that are used to map (convert) Dto to Entity
and vice versa.