POST request from the GuiClient was added
Now user, which is using GUI as a client, is able to click 'SignUp'
button, that action creates an event that is handled by RegistrationController's method -
onSignUp(). It start the process of sending POST HTTP request to the
server that creates new User entity.

All models, mappers and configuration class were deleted, because that
are useless in this app (at least now).