Receiving products from the server was added
The GUI client can send a request with POST HTTP method to the server
and receives all products as server's response. Products is showed in the products page.
ListView contains all of them.

MouseClicked event was added for each product in the ListView structure.
If user clicks on the product twice product's description as an information
alert opens.