Division into roles was added
If user has role USER he goes to the standard products page (products-page.fxml) after sign
in, but if he has role ADMIN he goes to the admin products page
(products-admin-page.fxml). The same thing with the order page.

A couple of abstract classes were added. OrderController and
ProductController are abstract classes that contain common fields and
methods for each Product and Order controller (for USER and ADMIN
roles).