Page navigation for user with role ADMIN was added
When user with role ADMIN sign in he goes to the main-admin-page.fxml.
There are some buttons that are used to move between admin's pages (for example,
products page -> products-admin-page.fxml).

The 'Back to main' button was added for each admin's page as well.
Administrator can't move to the orders-admin-page.fxml from the
products-admin-page.fxml anymore, he has to go to the
main-admin-page.fxml first and then choose needed page.