Sign out opportunity was added
When a user click secondary mouse button on his account information (the top
right corner in the pages) they sign out and move to the sign in page
(login-page.fxml).

DefaultUserAccountEventHandler is a component (Spring's bean) now that's
why it is not an utility class anymore.

The DELETE HTTP Method was added. RestRequestSender contains it and
ready to send requests to the REST server.